feat: adicionar validação de usuário na atualização e exclusão de usuários, além da atualização do método de update

+++ PomodoroAPI/Modules/Usuario/Controllers/UsuarioController.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using PomodoroAPI.Infrastructure.Services;
 using PomodoroAPI.Modules.Usuario.Models;
 using PomodoroAPI.Modules.Usuario.Repositories;
 
@@ -24,8 +26,9 @@ public class UsuarioController : ControllerBase
 
     [Authorize]
     [HttpPut("{id}")]
-    public async Task<ActionResult<UsuarioModel>> Update(int id, [FromBody] UsuarioModel usuarioBody)
+    public async Task<ActionResult<UsuarioModel>> Update(int id, [FromBody] UpdateUsuarioViewModel usuarioBody)
     {
+        if (AuthorizeService.GetUserId(User) != id) return Unauthorized();
         return Ok(await _usuarioRepository.Update(id, usuarioBody));
     }
 
@@ -33,6 +36,7 @@ public class UsuarioController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult> Delete(int id)
     {
+        if (AuthorizeService.GetUserId(User) != id) return Unauthorized();
         await _usuarioRepository.Delete(id);
         return NoContent();
     }
